+++ src/Slices/Resource/Core/Utils.ts
@@ -1,0 +1,7 @@
+import { RemoteData } from "@/Core";
+import { Manifest } from "./Query";
+
+export type ViewData = RemoteData.RemoteData<
+  Manifest["error"],
+  Manifest["usedData"]
+>;
