+++ src/Slices/Resource/Data/StateHelper.ts
@@ -2,7 +2,7 @@ import { RemoteData, Resource } from "@/Core";
 import { PrimaryStateHelperWithEnv } from "@/Data/Common";
 import { Store } from "@/Data/Store";
 
-export class ResourcesStateHelper extends PrimaryStateHelperWithEnv<"GetResources"> {
+export class GetResourcesStateHelper extends PrimaryStateHelperWithEnv<"GetResources"> {
   constructor(store: Store) {
     super(
       store,
