+++ src/Slices/Resource/Data/getUrl.ts
