+++ src/Slices/Resource/Data/index.ts
