+++ src/Slices/Resource/UI/ResourcesPage/Components/DeployButton.tsx
