+++ src/Slices/Resource/UI/ResourcesPage/Components/RequiresTableWithData.tsx
