+++ src/Slices/Resource/UI/ResourcesPage/Components/index.ts
