+++ src/Slices/Resource/UI/ResourcesPage/Page.tsx
