+++ src/Slices/Resource/UI/ResourcesPage/index.ts
